Adding a REPL.
start with 'node repl'

This repl has each individual model added to the context for easy access, and the controllers object added, just in case.  It's fully hooked up to the regular database, so queries, updates and saves all work as they would from the rails console.

The repl is promises-aware, so if the method you call returns a promise, the repl will wait until the promise has resolved, and print out that value. Alas, this means most of the Model methods can't be used in assignment statements, and you have to use 'suite = _' a lot.